simplify time series endpoint
