+++ backend/app/api_views.py
@@ -1,16 +1,9 @@
-import base64
 import json
-from io import BytesIO
 from pathlib import Path
 
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-
-# https://stackoverflow.com/questions/50157759/runtimeerror-main-thread-is-not-in-main-loop-using-matplotlib-with-django
-import matplotlib
-matplotlib.use('Agg')
-import matplotlib.pyplot as plt
 
 import numpy as np
 
@@ -18,13 +11,13 @@ from django.conf import settings
 
 from app.synthesis.audio_encoding import audio_samples_to_wav_base64
 from app.synthesis import synthesizers as synths
-from app.notes import NOTES
 
 from app.data_processing import (
     csv_files as csv_processing,
     color as color_processing,
     gesture as gesture_processing,
     polygon as polygon_processing,
+    time_series as time_series_processing,
     text as text_processing,
 )
 
@@ -159,125 +152,19 @@ def time_series_sample_data(request):
     return Response(csv_data)
 
 
-# pylint: disable=too-many-locals
 @api_view(['POST'])
-def time_series_to_music(request):
+def time_series_to_audio(request):
     """
-    Takes a dictionary representing a parsed CSV and constructs samples based on
-    those ratios.
+    Takes a dictionary representing a parsed CSV and constructs a piece of music 
+    and a set of samples based on the data.
     """
-    csv_data = request.data['parsedCSV']
-    column_constants = request.data['constants']
-    duration = float(request.data['duration'])
-    every_n = int(request.data['everyN'])
-    print('Music every n', every_n)
-    csv_data = csv_data[::every_n]
-    map_to_note = request.data['mapToNote']
+    music_data = time_series_processing.time_series_to_music(request)
+    samples = time_series_processing.time_series_to_samples(request)
 
-    new_csv = []
-    audio_samples = None
-
-    for i, row in enumerate(csv_data):
-        sound = None
-        new_csv_row = []
-        for j, frequency in enumerate(row):
-            if frequency == "":
-                new_csv_row += [0]
-                continue
-            column_constant = column_constants[j]
-
-            frequency =\
-                (float(frequency) + column_constant["offset"]) * column_constant["multiplier"]
-
-            if map_to_note:
-                # map number [0,88] to a note
-                frequency = (2 ** ((frequency - 49) / 12)) * 440
-
-            frequency += column_constant["base_frequency"]
-            new_csv_row += [frequency]
-
-            note = synths.generate_sine_wave_with_envelope(
-                frequency=frequency,
-                duration=duration,
-                a_percentage=column_constant["a_percentage"],
-                d_percentage=column_constant["d_percentage"],
-                s_percentage=column_constant["s_percentage"],
-                r_percentage=column_constant["r_percentage"],
-            )
-            if sound is None:
-                sound = note
-            else:
-                sound += note
-        new_csv += [new_csv_row]
-        if audio_samples is None:
-            audio_samples = sound
-        else:
-            audio_samples = np.append(audio_samples, sound)
-
-    sound = audio_samples_to_wav_base64(audio_samples)
-
-    time_steps = np.arange(0, len(csv_data))
-    new_csv = np.array(new_csv)
-    plt.plot(time_steps, new_csv)
-
-    min_f = np.amin(new_csv) - 10
-    max_f = np.amax(new_csv) + 10
-    for each in NOTES:
-        f = each["Frequency (Hz)"]
-        if min_f <= f <= max_f:
-            if "A" in each["Note"] and len(each["Note"]) == 2:
-                plt.axhline(y=f, color='r', linestyle='-')
-            else:
-                plt.axhline(y=f, color='grey', linestyle='-.')
-            plt.text(len(new_csv) - 1, f, each["Note"])
-
-    # TODO(ra) - let's add this back once it looks a bit nicer
-    # plt.legend(["Col " + str(i + 1) for i in range(len(new_csv[0]))])
-
-    plt.xlabel("Time Step")
-    plt.ylabel("Frequency (Hz)")
-    plt.yscale('log')
-
-    buffer = BytesIO()
-    plt.savefig(buffer, format='png')
-    buffer.seek(0)
-    img_str = base64.b64encode(buffer.getvalue())
-
-    plt.clf()
-
-    return Response({"sound": sound, "img": img_str})
-
-
-@api_view(['POST'])
-def time_series_to_samples(request):
-    """
-    Takes a dictionary representing a parsed CSV file and constructs samples
-    based on the column averages.
-    """
-    csv_data = request.data['parsedCSV']
-    column_constants = request.data['constants']
-
-    csv_np_array = np.array(csv_data).astype(np.float)
-
-    csv_row_av = np.average(csv_np_array, axis=0)
-    wav_files = []
-    for j, frequency in enumerate(csv_row_av):
-        column_constant = column_constants[j]
-        freq_to_generate = column_constant["base_frequency"] + (
-            float(frequency) + column_constant["offset"]) * column_constant[
-                               "multiplier"]
-        note = synths.generate_sine_wave_with_envelope(
-            frequency=freq_to_generate,
-            duration=1,
-            a_percentage=column_constant["a_percentage"],
-            d_percentage=column_constant["d_percentage"],
-            s_percentage=column_constant["s_percentage"],
-            r_percentage=column_constant["r_percentage"]
-        )
-        wav_file_base64 = audio_samples_to_wav_base64(note)
-        wav_files.append(wav_file_base64)
-
-    return Response(wav_files)
+    return Response({
+        'samples': samples,
+        'musicData': music_data,
+    })
 
 
 ################################################################################
